+++ PetShop.UI.API/Controllers/PetsController.cs
@@ -30,7 +30,11 @@ namespace PetShop.UI.API.Controllers
         {
             if (id < 1) return BadRequest("Id must be greater then 0");
             
-            return Ok(_petService.GetPet(id));
+            var pet = _petService.GetPet(id);
+
+            if (pet == null) return NotFound();
+            
+            return Ok(pet);
         }
         
         // POST api/pets -- CREATE
@@ -39,7 +43,8 @@ namespace PetShop.UI.API.Controllers
         {
             try
             {
-                return Ok(_petService.AddPet(pet));
+                var petFromDb = _petService.AddPet(pet);
+                return Created($"api/pets/{petFromDb.Id}", petFromDb);
             }
             catch (Exception e)
             {
@@ -51,20 +56,28 @@ namespace PetShop.UI.API.Controllers
         [HttpPut("{id}")]
         public ActionResult<Pet> Put(int id, [FromBody] Pet pet)
         {
+            if (id < 1) return BadRequest("Id must be greater then 0");
+            
             if (id < 1 || id != pet.Id)
             {
                 return BadRequest("Parameter Id and pet ID must be the same");
             }
+            
+            if (_petService.GetPet(id) == null) return NotFound();
 
-            return Ok(_petService.UpdatePet(pet));
+            return Accepted(_petService.UpdatePet(pet));
         }
         
         // DELETE api/pets/5
         [HttpDelete("{id}")]
         public ActionResult<Pet> Delete(int id)
         {
+            if (id < 1) return BadRequest("Id must be greater then 0");
+            
+            if (_petService.GetPet(id) == null) return NotFound();
+            
             _petService.DeletePet(id);
-            return Ok($"Pet with Id: {id} is Deleted");
+            return Accepted($"Pet with Id: {id} is Deleted");
         }
         
     }
