+++ PetShop.UI.API/Controllers/TypesController.cs
@@ -18,28 +18,34 @@ namespace PetShop.UI.API.Controllers
             _typeService = typeService; 
         }
             
+        // GET api/types
         [HttpGet]
         public IEnumerable<Type> Get()
         {
             return _typeService.GetTypes();
         }
         
-        // GET api/pettype/5 -- READ By Id
+        // GET api/types/5 -- READ By Id
         [HttpGet("{id}")]
         public ActionResult<Type> Get(int id)
         {
             if (id < 1) return BadRequest("Id must be greater then 0");
-            
-            return Ok(_typeService.GetType(id));
+
+            var type = _typeService.GetType(id);
+
+            if (type == null) return NotFound();
+
+            return Ok(type);
         }
         
-        // POST api/pettype -- CREATE
+        // POST api/types -- CREATE
         [HttpPost]
         public ActionResult<Type> Post([FromBody] Type type)
         {
             try
             {
-                return Ok(_typeService.AddType(type));
+                var typeFromDb = _typeService.AddType(type);
+                return Created($"api/types/{typeFromDb.Id}", typeFromDb);
             }
             catch (Exception e)
             {
@@ -47,24 +53,32 @@ namespace PetShop.UI.API.Controllers
             }
         }
         
-        // PUT api/pettype/5 -- Update
+        // PUT api/types/5 -- Update
         [HttpPut("{id}")]
         public ActionResult<Type> Put(int id, [FromBody] Type type)
         {
+            if (id < 1) return BadRequest("Id must be greater then 0");
+            
             if (id < 1 || id != type.Id)
             {
                 return BadRequest("Parameter Id and order ID must be the same");
             }
+            
+            if (_typeService.GetType(id) == null) return NotFound();
 
-            return Ok(_typeService.UpdateType(type));
+            return Accepted(_typeService.UpdateType(type));
         }
         
-        // DELETE api/pettype/5
+        // DELETE api/types/5
         [HttpDelete("{id}")]
         public ActionResult<Type> Delete(int id)
         {
+            if (id < 1) return BadRequest("Id must be greater then 0");
+            
+            if (_typeService.GetType(id) == null) return NotFound();
+            
             _typeService.DeleteType(id);
-            return Ok($"Type with Id: {id} is Deleted");
+            return Accepted($"Type with Id: {id} is Deleted");
         }
         
     }
